requirements.txt
# ensure you are using python <= 3.7 as well
numpy == 1.19.2
opencv-python == 4.5.1.48
opencv-utils == 0.0.2
tensorflow == 2.0.0
keras == 2.3.1
scikit-image == 0.17.2
flask-restful == 0.3.8
flask-cors == 3.0.10
redis == 3.5.3
h5py==2.10.0
Pillow